Made is_shadowed function slightly faster

+++ src/lighting/camera.c
@@ -6,7 +6,7 @@
 /*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/24 14:33:36 by mkhan             #+#    #+#             */
-/*   Updated: 2022/12/02 18:43:01 by hsarhan          ###   ########.fr       */
+/*   Updated: 2022/12/04 18:52:14 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,7 +48,7 @@ void	camera_init(t_camera *camera, t_scene *scene)
 	double	aspect;
 	
 
-	half_view = tanf((camera->fov / 2.0f) * M_PI / 180.0f);
+	half_view = tan((camera->fov / 2.0f) * M_PI / 180.0f);
 	aspect = scene->render_w / (double) scene->render_h;
 	if (aspect >= 1)
 	{
